Add OTLP logs exporter and move configuration to telemetry folder
