+++ src/shippingservice/src/telemetry/resources_conf.rs
@@ -1,0 +1,25 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+use opentelemetry_sdk::{
+    resource::{
+        EnvResourceDetector, OsResourceDetector, ProcessResourceDetector, ResourceDetector,
+        SdkProvidedResourceDetector, TelemetryResourceDetector,
+    },
+    Resource,
+};
+use std::time::Duration;
+
+pub fn get_resource_attr() -> Resource {
+    let os_resource = OsResourceDetector.detect(Duration::from_secs(0));
+    let process_resource = ProcessResourceDetector.detect(Duration::from_secs(0));
+    let sdk_resource = SdkProvidedResourceDetector.detect(Duration::from_secs(0));
+    let env_resource = EnvResourceDetector::new().detect(Duration::from_secs(0));
+    let telemetry_resource = TelemetryResourceDetector.detect(Duration::from_secs(0));
+
+    os_resource
+        .merge(&process_resource)
+        .merge(&sdk_resource)
+        .merge(&env_resource)
+        .merge(&telemetry_resource)
+}
